# Add SSE2 versions of ClampedAddSubtractFull and ClampedAddSubtractHalf