### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description

This PR adds SSE2 versions of ClampedAddSubtractFull and ClampedAddSubtractHalf which are used during encoding and decoding of lossless webp.

Related to: #1786

TODO:
- [x] Add tests.

Before:
![before](https://user-images.githubusercontent.com/38701097/139858554-035241f0-22fa-44e4-8cf2-2b12fb0a76e5.png)

After:
![After](https://user-images.githubusercontent.com/38701097/139858577-41915523-6b3c-49a6-a87e-f4da2cda218d.png)

Benchmark results encoding:
```
BenchmarkDotNet=v0.13.0, OS=Windows 10.0.19043.1320 (21H1/May2021Update)
Intel Core i7-6700K CPU 4.00GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET SDK=6.0.100-rc.2.21505.57
  [Host]     : .NET 5.0.11 (5.0.1121.47308), X64 RyuJIT
  Job-RZASZW : .NET 5.0.11 (5.0.1121.47308), X64 RyuJIT
  Job-RSRLCF : .NET Core 3.1.20 (CoreCLR 4.700.21.47003, CoreFX 4.700.21.47101), X64 RyuJIT
  Job-JFGCRR : .NET Framework 4.8 (4.8.4420.0), X64 RyuJIT

IterationCount=3  LaunchCount=1  WarmupCount=3

|                     Method |        Job |              Runtime |             Arguments |    TestImage |      Mean |     Error |   StdDev | Ratio | RatioSD |       Gen 0 |     Gen 1 |     Gen 2 |  Allocated |
|--------------------------- |----------- |--------------------- |---------------------- |------------- |----------:|----------:|---------:|------:|--------:|------------:|----------:|----------:|-----------:|
|        'Magick Webp Lossy' | Job-RZASZW |             .NET 5.0 | /p:DebugType=portable | Png/Bike.png |  23.36 ms |  2.292 ms | 0.126 ms |  0.14 |    0.00 |           - |         - |         - |      68 KB |
|    'ImageSharp Webp Lossy' | Job-RZASZW |             .NET 5.0 | /p:DebugType=portable | Png/Bike.png | 247.26 ms | 29.033 ms | 1.591 ms |  1.53 |    0.00 | 135000.0000 |         - |         - | 552,714 KB |
|     'Magick Webp Lossless' | Job-RZASZW |             .NET 5.0 | /p:DebugType=portable | Png/Bike.png | 161.25 ms | 11.824 ms | 0.648 ms |  1.00 |    0.00 |           - |         - |         - |     519 KB |
| 'ImageSharp Webp Lossless' | Job-RZASZW |             .NET 5.0 | /p:DebugType=portable | Png/Bike.png | 305.26 ms | 66.906 ms | 3.667 ms |  1.89 |    0.03 |  34000.0000 | 5000.0000 | 2000.0000 | 161,674 KB |
|                            |            |                      |                       |              |           |           |          |       |         |             |           |           |            |
|        'Magick Webp Lossy' | Job-RSRLCF |        .NET Core 3.1 |               Default | Png/Bike.png |  23.38 ms |  0.190 ms | 0.010 ms |  0.14 |    0.00 |           - |         - |         - |      68 KB |
|    'ImageSharp Webp Lossy' | Job-RSRLCF |        .NET Core 3.1 |               Default | Png/Bike.png | 256.32 ms | 36.712 ms | 2.012 ms |  1.59 |    0.01 | 135000.0000 |         - |         - | 552,713 KB |
|     'Magick Webp Lossless' | Job-RSRLCF |        .NET Core 3.1 |               Default | Png/Bike.png | 161.41 ms |  8.593 ms | 0.471 ms |  1.00 |    0.00 |           - |         - |         - |     523 KB |
| 'ImageSharp Webp Lossless' | Job-RSRLCF |        .NET Core 3.1 |               Default | Png/Bike.png | 321.77 ms | 30.517 ms | 1.673 ms |  1.99 |    0.01 |  34000.0000 | 5000.0000 | 2000.0000 | 161,673 KB |
|                            |            |                      |                       |              |           |           |          |       |         |             |           |           |            |
|        'Magick Webp Lossy' | Job-JFGCRR | .NET Framework 4.7.2 |               Default | Png/Bike.png |  23.43 ms |  1.900 ms | 0.104 ms |  0.14 |    0.00 |           - |         - |         - |      68 KB |
|    'ImageSharp Webp Lossy' | Job-JFGCRR | .NET Framework 4.7.2 |               Default | Png/Bike.png | 377.71 ms | 34.625 ms | 1.898 ms |  2.33 |    0.02 | 135000.0000 |         - |         - | 554,351 KB |
|     'Magick Webp Lossless' | Job-JFGCRR | .NET Framework 4.7.2 |               Default | Png/Bike.png | 161.85 ms | 13.271 ms | 0.727 ms |  1.00 |    0.00 |           - |         - |         - |     520 KB |
| 'ImageSharp Webp Lossless' | Job-JFGCRR | .NET Framework 4.7.2 |               Default | Png/Bike.png | 387.72 ms | 14.809 ms | 0.812 ms |  2.40 |    0.01 |  34000.0000 | 5000.0000 | 2000.0000 | 162,118 KB |
```

Benchmark results decoding:
```
|                     Method |        Job |              Runtime |             Arguments |        TestImageLossy |        TestImageLossless |       Mean |       Error |   StdDev |    Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------- |----------- |--------------------- |---------------------- |---------------------- |------------------------- |-----------:|------------:|---------:|---------:|------:|------:|----------:|
|        'Magick Lossy Webp' | Job-RXAKRB |             .NET 5.0 | /p:DebugType=portable | Webp/earth_lossy.webp | Webp/earth_lossless.webp |   111.0 ms |    40.93 ms |  2.24 ms |        - |     - |     - |     25 KB |
|    'ImageSharp Lossy Webp' | Job-RXAKRB |             .NET 5.0 | /p:DebugType=portable | Webp/earth_lossy.webp | Webp/earth_lossless.webp |   286.2 ms |   108.16 ms |  5.93 ms |        - |     - |     - |  2,428 KB |
|     'Magick Lossless Webp' | Job-RXAKRB |             .NET 5.0 | /p:DebugType=portable | Webp/earth_lossy.webp | Webp/earth_lossless.webp |   115.0 ms |    72.28 ms |  3.96 ms |        - |     - |     - |     16 KB |
| 'ImageSharp Lossless Webp' | Job-RXAKRB |             .NET 5.0 | /p:DebugType=portable | Webp/earth_lossy.webp | Webp/earth_lossless.webp |   311.2 ms |   529.47 ms | 29.02 ms |        - |     - |     - |  2,091 KB |
|        'Magick Lossy Webp' | Job-SHZFXZ |        .NET Core 3.1 |               Default | Webp/earth_lossy.webp | Webp/earth_lossless.webp |   124.1 ms |    74.47 ms |  4.08 ms |        - |     - |     - |     25 KB |
|    'ImageSharp Lossy Webp' | Job-SHZFXZ |        .NET Core 3.1 |               Default | Webp/earth_lossy.webp | Webp/earth_lossless.webp |   308.8 ms |   271.16 ms | 14.86 ms | 500.0000 |     - |     - |  2,428 KB |
|     'Magick Lossless Webp' | Job-SHZFXZ |        .NET Core 3.1 |               Default | Webp/earth_lossy.webp | Webp/earth_lossless.webp |   112.0 ms |    43.24 ms |  2.37 ms |        - |     - |     - |     15 KB |
| 'ImageSharp Lossless Webp' | Job-SHZFXZ |        .NET Core 3.1 |               Default | Webp/earth_lossy.webp | Webp/earth_lossless.webp |   462.8 ms |   188.74 ms | 10.35 ms |        - |     - |     - |  2,092 KB |
|        'Magick Lossy Webp' | Job-VLFICQ | .NET Framework 4.7.2 |               Default | Webp/earth_lossy.webp | Webp/earth_lossless.webp |   108.5 ms |    66.75 ms |  3.66 ms |        - |     - |     - |     32 KB |
|    'ImageSharp Lossy Webp' | Job-VLFICQ | .NET Framework 4.7.2 |               Default | Webp/earth_lossy.webp | Webp/earth_lossless.webp |   582.0 ms |   169.32 ms |  9.28 ms |        - |     - |     - |  2,436 KB |
|     'Magick Lossless Webp' | Job-VLFICQ | .NET Framework 4.7.2 |               Default | Webp/earth_lossy.webp | Webp/earth_lossless.webp |   109.7 ms |    28.50 ms |  1.56 ms |        - |     - |     - |     18 KB |
| 'ImageSharp Lossless Webp' | Job-VLFICQ | .NET Framework 4.7.2 |               Default | Webp/earth_lossy.webp | Webp/earth_lossless.webp | 1,865.8 ms | 1,250.04 ms | 68.52 ms |        - |     - |     - |  9,729 KB |
```